puaheo #2 creacion de modelo usuario funcional ruta post de usuaruio conectar componente comentpage a create coment

+++ back/src/Controllers/EliminarComentarios.js
@@ -2,10 +2,20 @@ const { Comentario } = require('../db');
 
 const EliminarComentario = async (id) => { 
   try {
-    const eliminar = await Comentario.destroy({ where: { id } });
-    return 'comentario eliminado exitosamente';
+    if (!id) {
+      throw new Error('El ID del comentario es inválido');
+    }
+
+    const comentarioEliminado = await Comentario.destroy({ where: { id } });
+
+    if (comentarioEliminado === 0) {
+      throw new Error('No se encontró ningún comentario con el ID proporcionado');
+    }
+
+    return 'Comentario eliminado exitosamente';
   } catch (error) {
-    throw new Error(error.message); 
+    console.error('Error al eliminar el comentario:', error);
+    throw new Error('No se pudo eliminar el comentario');
   }
 };
